route.tsx 파일 추가/자동 로그인
